remove clusters, keep only monorepo one
