--- packages/symfony-static-dump/src/ControllerWithDataProvider/ClusterControllerWithDataProvider.php
@@ -1,41 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace TomasVotruba\SymfonyStaticDump\ControllerWithDataProvider;
-
-use Symplify\SymfonyStaticDumper\Contract\ControllerWithDataProviderInterface;
-use TomasVotruba\Blog\Repository\ClusterRepository;
-use TomasVotruba\Website\Controller\ClusterController;
-
-final class ClusterControllerWithDataProvider implements ControllerWithDataProviderInterface
-{
-    public function __construct(
-        private ClusterRepository $clusterRepository
-    ) {
-    }
-
-    public function getControllerClass(): string
-    {
-        return ClusterController::class;
-    }
-
-    public function getControllerMethod(): string
-    {
-        return '__invoke';
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getArguments(): array
-    {
-        $slugs = [];
-
-        foreach ($this->clusterRepository->getClusters() as $cluster) {
-            $slugs[] = $cluster->getSlug();
-        }
-
-        return $slugs;
-    }
-}
