+++ templates/contact.twig
@@ -1,17 +1,14 @@
 {% extends "base.twig" %}
 
 {% block content %}
-    <div class="container" id="contact">
+    <div class="container">
         <h1>{{ title }}</h1>
 
-        <p class="text-center bigger">
-            Do you need my services? We can start today:
-        </p>
+        <p class="text-center bigger">What is the best way to reach me?</p>
 
-        <div class="row mt-4 mb-5 ml-sm-0 ml-md-5 " id="contact-points">
-
+        <div class="row mt-4 mb-5 ml-sm-0 ml-md-5">
             <div class="col-8 pt-md-3 pt-lg-3 pl-md-4 pl-lg-5">
-                The best way to reach me is direct real-time call: <a href="tel:+420776778332">+420 <strong>776 778 332</strong></a>. If you won't hear from me till next day, send me SMS.
+                The best way to reach me is direct real-time call: <a href="tel:+420776778332">+420 <strong>776 778 332</strong></a>. If you won't hear from me till next day, send me SMS and I'll call you.
 
                 When? Workdays between 2-6 PM.
